[JSC] Implement RegExp.prototype[Symbol.replace] in C++ https://bugs.webkit.org/show_bug.cgi?id=304752 rdar://167280055
Reviewed by Sosuke Suzuki.

This patch implements RegExp.prototype[Symbol.replace] in C++ instead of
JS. This is not super important function since String.prototype.replace
and its fast path is mostly called, and this function is rarely called.
Let's just have it in C++ so that we can simplify our builtin
implementations more.
We generate bunch of tests for RegExp.prototype[Symbol.replace] to cover
the cases.

Tests: JSTests/stress/regexp-prototype-symbol-replace-basic.js
       JSTests/stress/regexp-prototype-symbol-replace-capture-groups.js
       JSTests/stress/regexp-prototype-symbol-replace-edge-cases.js
       JSTests/stress/regexp-prototype-symbol-replace-functional.js
       JSTests/stress/regexp-prototype-symbol-replace-named-groups.js
       JSTests/stress/regexp-prototype-symbol-replace-slow-path.js
       JSTests/stress/regexp-prototype-symbol-replace-unicode.js

* JSTests/stress/regexp-prototype-symbol-replace-basic.js: Added.
(shouldBe):
* JSTests/stress/regexp-prototype-symbol-replace-capture-groups.js: Added.
(shouldBe):
* JSTests/stress/regexp-prototype-symbol-replace-edge-cases.js: Added.
(shouldBe):
(shouldThrow):
(toString):
(let.fakeRegExp.exec):
* JSTests/stress/regexp-prototype-symbol-replace-functional.js: Added.
(shouldBe):
(return.toString):
* JSTests/stress/regexp-prototype-symbol-replace-named-groups.js: Added.
(shouldBe):
* JSTests/stress/regexp-prototype-symbol-replace-slow-path.js: Added.
(shouldBe):
(reWithCustomExec.exec):
(customResult.exec):
(customLength.exec):
(modifyingExec.exec):
(nullExec.exec):
(let.plainObject.exec):
(throwingExec.exec):
(i.i.2.0.re.exec):
(funcWithCustomExec.exec):
(slowPathRe.exec):
(namedCapturesExec.exec):
* JSTests/stress/regexp-prototype-symbol-replace-unicode.js: Added.
(shouldBe):
(catch):
* Source/JavaScriptCore/builtins/RegExpPrototype.js:
(linkTimeConstant.getSubstitution): Deleted.
(overriddenName.string_appeared_here.replace): Deleted.
* Source/JavaScriptCore/dfg/DFGFixupPhase.cpp:
(JSC::DFG::FixupPhase::addStringReplacePrimordialChecks):
* Source/JavaScriptCore/runtime/RegExpObjectInlines.h:
(JSC::advanceStringUnicode):
* Source/JavaScriptCore/runtime/RegExpPrototype.cpp:
(JSC::RegExpPrototype::finishCreation):
(JSC::advanceStringIndex):
(JSC::getSubstitution):
(JSC::JSC_DEFINE_HOST_FUNCTION):

Canonical link: https://commits.webkit.org/304983@main

+++ Source/JavaScriptCore/runtime/RegExpObjectInlines.h
@@ -155,7 +155,7 @@ ALWAYS_INLINE MatchResult RegExpObject::matchInline(JSGlobalObject* globalObject
     return result;
 }
 
-inline unsigned advanceStringUnicode(StringView s, unsigned length, unsigned currentIndex)
+inline uint64_t advanceStringUnicode(StringView s, unsigned length, uint64_t currentIndex)
 {
     if (currentIndex + 1 >= length)
         return currentIndex + 1;
